+++ api/index.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
+
 import { createClient } from 'contentful'
 
 export const client = createClient({
