+++ app/[locale]/layout.tsx
@@ -5,12 +5,9 @@ import { type PropsWithChildren } from 'react'
 import { type Metadata } from 'next'
 import { notFound } from 'next/navigation'
 
-import GoogleAnalytics from '@/components/GoogleAnalytics'
-import Header from '@/components/Header'
-import LanguagePicker from '@/components/LanguagePicker'
-import metadata from '@/constants/metadata'
-import { locales } from '@/routes/locales'
-import { navigation } from '@/routes/navigation'
+import { GoogleAnalytics, Header, LanguagePicker } from '@/components'
+import { metadata } from '@/constants'
+import { locales, navigation } from '@/routes'
 
 import '@mantine/core/styles.css'
 import '@mantine/core/styles/global.css'
