+++ constants/index.ts
@@ -1,0 +1,1 @@
+export { default as metadata } from './metadata'
