+++ helpers/index.ts
@@ -1,0 +1,1 @@
+export { GA_TRACKING_ID, event, pageview } from './gtag'
