+++ modules/index.ts
@@ -1,0 +1,2 @@
+export { default as Homepage } from './Homepage'
+export { default as PostDetail } from './PostDetail'
