+++ routes/index.ts
@@ -1,0 +1,3 @@
+export { default as locales } from './locales'
+export { default as navigation } from './navigation'
+export { default as urls } from './urls'
